refactor(03-01): introduce ISqlDialect strategy interface and *Dialect classes
- Add ISqlDialect with 4 methods: GetDueJobs, ReleaseLeasedJobs, GetDueSchedules, UpsertScheduleAsync
- Create PostgreSqlDialect, SqlServerDialect, MySqlDialect as internal sealed implementations
- Copy SQL verbatim from *Provider classes; rename methods to drop misleading Async suffix
- Add UpsertScheduleAsync stub (NotImplementedException, TODO: Implemented in Phase 4)
- Delete IDatabaseProviderSql, PostgreSqlProvider, SqlServerProvider, MySqlProvider
