+++ src/Atomizer.EntityFrameworkCore/Providers/ISqlDialect.cs
@@ -1,0 +1,9 @@
+namespace Atomizer.EntityFrameworkCore.Providers;
+
+internal interface ISqlDialect
+{
+    FormattableString GetDueJobs(QueueKey queueKey, DateTimeOffset now, int batchSize);
+    FormattableString ReleaseLeasedJobs(LeaseToken leaseToken, DateTimeOffset now);
+    FormattableString GetDueSchedules(DateTimeOffset now);
+    FormattableString UpsertScheduleAsync(AtomizerSchedule schedule);
+}
